feat: add MIT license, enhance README, and implement 'Why Boycott' page with navigation updates
