+++ src/app/about/page.tsx
@@ -1,7 +1,7 @@
 import { MarkdownPage } from "~/components/MarkdownPage";
 
 export const metadata = {
-	title: "About | Boycott Israel",
+	title: "About Boycott Israel BD",
 	description: "Learn about our mission, team, and frequently asked questions.",
 };
