+++ src/app/why-boycott/page.tsx
@@ -1,11 +1,11 @@
 import { MarkdownPage } from "~/components/MarkdownPage";
 
 export const metadata = {
-	title: "Resources | Boycott Israel",
+	title: "Why Boycott Israel",
 	description:
 		"Educational resources about the boycott movement, guidelines, and news updates.",
 };
 
 export default function ResourcesPage() {
-	return <MarkdownPage filePath="src/content/resources.mdx" />;
+	return <MarkdownPage filePath="src/content/why-boycott.mdx" />;
 }
