+++ src/content/why-boycott.mdx
@@ -1,4 +1,4 @@
-# Resources
+# Why Boycottt?
 
 ## Boycott Movement
